requirement.txt
pytest>=7.0
requests>=2.28
pyyaml
jsonschema